ml_backend/requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Model Export & Inference
onnx>=1.14.0
onnxruntime>=1.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging & Monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
